ref: address review feedback — extract helper, consolidate tests
- Extract checkParameterEnumDiff helper to deduplicate top-level and
  property enum logic, following the check_response_status_updated.go
  pattern
- Merge property enum tests into check_request_parameter_enum_value_updated_test.go
  to mirror source file naming convention
- Add comment explaining the swap base/revision technique in the
  added-value test
